Adjust the response window for multiple tabs
This replaces the old mechanism that kept track of the response window
state by updating a variable representing it's visibility. Discovering
if the buffer is visible is now done by looking up the buffers of the
current tab.

A new functionallity is that with this change a repeated call to
IdrisResponseWin (likely through <LocalLeader>i) closes the window
again, making it very easy to open it, check some output, then close
it as needed to free up screen space.